# apps/openapi.yaml
openapi: 3.1.0
info:
  title: Quant Trading System API
  version: "1.2.0"
  description: >
    OpenAPI specification for the current quant-trading-system project.
    It is organized around the real backend domain model:
    Strategy -> StrategyRun -> Signal / Transaction / PortfolioSnapshot,
    plus market-data and broker-trading integrations.

    Notes:
    - Operations marked with x-implementation-status: implemented are already reflected by the current FastAPI codebase.
    - Operations marked with x-implementation-status: planned are aligned with existing service-layer and database structure,
      and are intended to guide the next API/router implementation pass.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/paper-trading/run:
    post:
      tags: [Paper Trading]
      summary: Run one paper strategy
      operationId: runPaperStrategy
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PaperTradingRunRequest"}
      responses:
        "201":
          description: Paper run completed; support/resistance BUYs may await the next open
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PaperTradingRunOut"}
  /api/paper-trading/run-multi:
    post:
      tags: [Paper Trading]
      summary: Run allocated paper strategies
      operationId: runPaperStrategies
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MultiStrategyPaperTradingRunRequest"}
      responses:
        "201":
          description: Portfolio paper-run results
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MultiStrategyPaperTradingRunOut"}
...
components:
  securitySchemes:
    AgentServiceToken:
      type: http
      scheme: bearer
      bearerFormat: opaque service token
...
  schemas:
    PaperTradingRunRequest:
      type: object
      required: [strategy_id, trade_date, portfolio_name]
      properties:
        strategy_id: {type: string, format: uuid}
        trade_date: {type: string, format: date}
        portfolio_name: {type: string, default: default}
        basket_id: {type: string, format: uuid, nullable: true}
        submit_orders: {type: boolean, default: true}
    MultiStrategyPaperTradingRunRequest:
      type: object
      required: [trade_date, portfolio_name]
      properties:
        trade_date: {type: string, format: date}
        portfolio_name: {type: string, default: default}
        submit_orders: {type: boolean, default: true}
        continue_on_error: {type: boolean, default: false}
    PaperTradingRunOut:
      type: object
      required: [run_id, strategy_id, status, trade_date, portfolio_name, allocation_pct,
        capital_base, signal_count, order_count, submitted_order_count, skipped_order_count,
        failed_order_count, pending_order_count, final_cash, final_equity]
      properties:
        run_id: {type: string, format: uuid}
        strategy_id: {type: string, format: uuid}
        status: {type: string}
        trade_date: {type: string, format: date}
        portfolio_name: {type: string}
        allocation_pct: {type: number}
        capital_base: {type: number}
        signal_count: {type: integer, minimum: 0}
        order_count: {type: integer, minimum: 0}
        submitted_order_count: {type: integer, minimum: 0}
        skipped_order_count: {type: integer, minimum: 0}
        failed_order_count: {type: integer, minimum: 0}
        pending_order_count:
          type: integer
          minimum: 0
          description: Support/resistance BUY intents awaiting strict next-open validation.
        final_cash: {type: number}
        final_equity: {type: number}
    MultiStrategyPaperTradingRunOut:
      type: object
      required: [portfolio_name, trade_date, total_runs, completed_runs, failed_runs, results]
      properties:
        portfolio_name: {type: string}
        trade_date: {type: string, format: date}
        total_runs: {type: integer, minimum: 0}
        completed_runs: {type: integer, minimum: 0}
        failed_runs: {type: integer, minimum: 0}
        results:
          type: array
          items: {$ref: "#/components/schemas/PaperTradingRunOut"}
...